Add some starting layouts for building final site
The file `layout.erb` is based on the existing Hugo build of my static
site, although it is missing a bunch of stuff like the social links.
Something got broken somewhere along the way and I'm not sure how.

The file `layout2.erb` is based on one of the example sites from the
purecss site. It can be found here:
https://pure-css.github.io/layouts/blog/

I'll probably just use the purecss one and pull stuff in from the Hugo
based one as needed.